+++ packages/@aws-c2a/broadening-permissions/.eslintignore
@@ -1,0 +1,2 @@
+out
+*.d.ts
