+++ packages/@aws-c2a/broadening-permissions/.eslintrc.js
@@ -1,0 +1,3 @@
+const baseConfig = require('../../../.eslintrc.build');
+baseConfig.parserOptions.project = __dirname + '/tsconfig.json';
+module.exports = baseConfig;
