+++ packages/@aws-c2a/broadening-permissions/babel.config.js
@@ -1,0 +1,7 @@
+// eslint-disable-next-line no-undef
+module.exports = {
+    presets: [
+        ['@babel/preset-env', {targets: {node: 'current'}}],
+        '@babel/preset-typescript',
+    ],
+};
