+++ packages/@aws-c2a/broadening-permissions/jest.config.js
@@ -1,5 +1,6 @@
 const base = require('../../../jest.config');
 module.exports = {
   ...base,
-  rootDir: './'
+  rootDir: './',
+  setupFilesAfterEnv: ['./jest.js'],
 };
