+++ packages/@aws-c2a/broadening-permissions/lib/index.ts
@@ -1,0 +1,1 @@
+export * from './security-changes';
