+++ packages/@aws-c2a/broadening-permissions/lib/private/object.ts
