+++ packages/@aws-c2a/broadening-permissions/lib/private/security-policies.ts
