+++ packages/@aws-c2a/broadening-permissions/test/iam-policy-rules.test.ts
