+++ packages/@aws-c2a/rules/lib/rules/index.ts
@@ -3,4 +3,3 @@ export * from './change';
 export * from './component';
 export * from './condition';
 export * from './rule';
-export * from './security-changes';
